fix(swiper): demo拆解与规范
