+++ src/packages/swiper/demo.taro.tsx
@@ -1,30 +1,23 @@
-import React, { useState, useEffect, useRef } from 'react'
+import React from 'react'
 import '@/packages/swiper/demo.scss'
 import Taro from '@tarojs/taro'
-import { ArrowLeft, ArrowRight } from '@nutui/icons-react-taro'
 import { useTranslate } from '@/sites/assets/locale/taro'
-import { Swiper } from '@/packages/nutui.react.taro'
 import Header from '@/sites/components/header'
-
-interface T {
-  basic: string
-  asyc: string
-  dynamicDel: string
-  size: string
-  indicator: string
-  btns: string
-  vertical: string
-  horizontalCenter: string
-  verticalCenter: string
-  multiItems: string
-}
+import Demo1 from './demos/taro/demo1'
+import Demo2 from './demos/taro/demo2'
+import Demo3 from './demos/taro/demo3'
+import Demo4 from './demos/taro/demo4'
+import Demo5 from './demos/taro/demo5'
+import Demo6 from './demos/taro/demo6'
+import Demo7 from './demos/taro/demo7'
+import Demo8 from './demos/taro/demo8'
+import Demo9 from './demos/taro/demo9'
 
 const SwiperDemo = () => {
-  const [translated] = useTranslate<T>({
+  const [translated] = useTranslate({
     'zh-CN': {
       basic: '基础用法',
       asyc: '异步加载(3s)',
-      dynamicDel: '动态加载',
       size: '自定义大小',
       indicator: '自定义指示器',
       btns: '手动切换',
@@ -36,7 +29,6 @@ const SwiperDemo = () => {
     'en-US': {
       basic: 'Basic Usage',
       asyc: 'Asynchronous loading(3s)',
-      dynamicDel: 'Dynamic loading',
       size: 'Custom size',
       indicator: 'Custom indicator',
       btns: 'Manual switching',
@@ -47,198 +39,28 @@ const SwiperDemo = () => {
     },
   })
 
-  const list = [
-    'https://storage.360buyimg.com/jdc-article/NutUItaro34.jpg',
-    'https://storage.360buyimg.com/jdc-article/NutUItaro2.jpg',
-    'https://storage.360buyimg.com/jdc-article/welcomenutui.jpg',
-    'https://storage.360buyimg.com/jdc-article/fristfabu.jpg',
-  ]
-
-  const [asyncList, setAsyncList] = useState<string[]>([])
-  useEffect(() => {
-    setTimeout(() => {
-      setAsyncList(list)
-    }, 3000)
-  }, [])
-
-  const [current, setCurrent] = useState(0)
-  const [current2, setCurrent2] = useState(0)
-  const swiperRef = useRef<any>(null)
-
   return (
     <>
       <Header />
       <div className={`demo ${Taro.getEnv() === 'WEB' ? 'web' : ''} padding`}>
         <h2>{translated.basic}</h2>
-        <Swiper defaultValue={1} autoPlay indicator>
-          {list.map((item, index) => (
-            <Swiper.Item key={item}>
-              <img
-                width="100%"
-                height="100%"
-                onClick={() => console.log(index)}
-                src={item}
-                alt=""
-              />
-            </Swiper.Item>
-          ))}
-        </Swiper>
-
+        <Demo1 />
         <h2>{translated.asyc}</h2>
-        <Swiper defaultValue={0} indicator>
-          {asyncList.map((item, index) => (
-            <Swiper.Item key={item}>
-              <img width="100%" height="100%" src={item} alt="" />
-            </Swiper.Item>
-          ))}
-        </Swiper>
-
+        <Demo2 />
         <h2>{translated.size}</h2>
-        <Swiper
-          className="swiper-demo-size"
-          width={300}
-          height={150}
-          defaultValue={0}
-          indicator
-        >
-          {list.map((item) => (
-            <Swiper.Item key={item}>
-              <img width="100%" height="100%" src={item} alt="" />
-            </Swiper.Item>
-          ))}
-        </Swiper>
-
+        <Demo3 />
         <h2>{translated.indicator}</h2>
-        <Swiper
-          defaultValue={0}
-          onChange={(e) => {
-            console.log('e.detail.current', e.detail.current)
-            setCurrent(e.detail.current)
-          }}
-          indicator={
-            <div
-              style={{
-                position: 'absolute',
-                bottom: '0',
-                right: '0',
-                width: '46px',
-                height: '22px',
-                background: 'rgba(0, 0, 0, 0.33)',
-                borderRadius: '22px',
-                textAlign: 'center',
-                color: '#fff',
-                fontSize: '14px',
-                zIndex: '1',
-              }}
-            >
-              {current + 1}/{list.length}
-            </div>
-          }
-        >
-          {list.map((item) => (
-            <Swiper.Item key={item}>
-              <img width="100%" height="100%" src={item} alt="" />
-            </Swiper.Item>
-          ))}
-        </Swiper>
-
+        <Demo4 />
         <h2>{translated.btns}</h2>
-        <div className="demo-box" style={{ height: 150 }}>
-          <Swiper
-            ref={swiperRef}
-            defaultValue={0}
-            onChange={(e) => {
-              setCurrent2(e.detail.current)
-            }}
-            indicator={
-              <div className="page">
-                {current2 + 1}/{list.length}
-              </div>
-            }
-          >
-            {list.map((item) => {
-              return (
-                <Swiper.Item key={item}>
-                  <img src={item} alt="" />
-                </Swiper.Item>
-              )
-            })}
-          </Swiper>
-          <div className="nut-swiper-btns">
-            <span
-              className="nut-swiper-btns-left"
-              onClick={(e) => swiperRef.current?.prev()}
-            >
-              <ArrowLeft />
-            </span>
-            <span
-              className="nut-swiper-btns-left"
-              onClick={(e) => swiperRef.current?.next()}
-            >
-              <ArrowRight />
-            </span>
-          </div>
-        </div>
+        <Demo5 />
         <h2>{translated.vertical}</h2>
-        <Swiper defaultValue={0} direction="vertical" indicator>
-          {list.map((item) => (
-            <Swiper.Item key={item}>
-              <img width="100%" height="100%" src={item} alt="" />
-            </Swiper.Item>
-          ))}
-        </Swiper>
+        <Demo6 />
         <h2>{translated.horizontalCenter}</h2>
-        <Swiper defaultValue={0} loop previousMargin="20px" nextMargin="20px">
-          {list.map((item) => (
-            <Swiper.Item key={item}>
-              <img width="100%" height="100%" src={item} alt="" />
-            </Swiper.Item>
-          ))}
-        </Swiper>
+        <Demo7 />
         <h2>{translated.verticalCenter}</h2>
-        <div className="demo-box vertical-center">
-          <Swiper
-            defaultValue={0}
-            direction="vertical"
-            height={220}
-            previousMargin="20px"
-            nextMargin="20px"
-            indicator
-          >
-            {list.map((item) => (
-              <Swiper.Item key={item}>
-                <img width="100%" height="100%" src={item} alt="" />
-              </Swiper.Item>
-            ))}
-          </Swiper>
-        </div>
+        <Demo8 />
         <h2>{translated.multiItems}</h2>
-        <Swiper loop autoPlay>
-          <Swiper.Item>
-            <div style={{ display: 'flex' }}>
-              <div style={{ flex: '1', border: '1 red' }}>Item1</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item2</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item3</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item4</div>
-            </div>
-          </Swiper.Item>
-          <Swiper.Item>
-            <div style={{ display: 'flex' }}>
-              <div style={{ flex: '1', border: '1 red' }}>Item5</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item6</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item7</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item8</div>
-            </div>
-          </Swiper.Item>
-          <Swiper.Item>
-            <div style={{ display: 'flex' }}>
-              <div style={{ flex: '1', border: '1 red' }}>Item9</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item10</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item11</div>
-              <div style={{ flex: '1', border: '1 red' }}>Item12</div>
-            </div>
-          </Swiper.Item>
-        </Swiper>
+        <Demo9 />
       </div>
     </>
   )
